requirements.txt
django==3.1.4
pandas
requests
time
djangorestframework